requirements.txt
# Data Processing
pandas>=2.0.0

# Machine Learning
scikit-learn>=1.3.0

# Data Visualization
matplotlib>=3.7.0

# HTTP Requests
requests>=2.31.0

# Environment Variables
python-dotenv>=1.0.0